Move pnpm catalog extraction to npm manager
