+++ lib/modules/manager/npm/extract/types.ts
@@ -36,6 +36,8 @@ export interface LockFile {
 
 export interface PnpmWorkspaceFile {
   packages: string[];
+  catalog?: Partial<Record<string, string>>;
+  catalogs?: Partial<Record<string, Partial<Record<string, string>>>>;
 }
 
 export type OverrideDependency = Record<string, RecursiveOverride>;
