--- lib/modules/manager/pnpm-catalog/index.ts
@@ -1,24 +1,0 @@
-import type { Category } from '../../../constants';
-import { GithubTagsDatasource } from '../../datasource/github-tags';
-import { NpmDatasource } from '../../datasource/npm';
-
-export { extractAllPackageFiles } from './extract';
-export { getRangeStrategy, updateDependency } from '../npm';
-
-export const categories: Category[] = ['js'];
-
-export const defaultConfig = {
-  fileMatch: ['(^|/)pnpm-workspace\\.yaml$'],
-  digest: {
-    prBodyDefinitions: {
-      Change:
-        '{{#if displayFrom}}`{{{displayFrom}}}` -> {{else}}{{#if currentValue}}`{{{currentValue}}}` -> {{/if}}{{/if}}{{#if displayTo}}`{{{displayTo}}}`{{else}}`{{{newValue}}}`{{/if}}',
-    },
-  },
-  prBodyDefinitions: {
-    Change:
-      "[{{#if displayFrom}}`{{{displayFrom}}}` -> {{else}}{{#if currentValue}}`{{{currentValue}}}` -> {{/if}}{{/if}}{{#if displayTo}}`{{{displayTo}}}`{{else}}`{{{newValue}}}`{{/if}}]({{#if depName}}https://renovatebot.com/diffs/npm/{{replace '/' '%2f' depName}}/{{{currentVersion}}}/{{{newVersion}}}{{/if}})",
-  },
-};
-
-export const supportedDatasources = [GithubTagsDatasource.id, NpmDatasource.id];
