--- lib/modules/manager/pnpm-catalog/readme.md
@@ -1,4 +1,0 @@
-Used for updating bun projects.
-Bun is a tool for JavaScript projects and therefore an alternative to managers like npm, pnpm and Yarn.
-
-If a `package.json` is found to be part of `bun` manager results then the same file will be excluded from the `npm` manager results unless an npm/pnpm/Yarn lock file is also found.
